SSDI_GRP_04/src/main/java/com/spring/app/bean/ContactBean.java
package com.spring.app.bean;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name="contact")
public class ContactBean {
	@Id
	@Column(name="email")
	String email;
	@Column(name="name")
	String name;
	@Column(name="comment")
	String Comment;

	public String getEmail() {
		return email;
	}

	public void setEmail(String email) {
		this.email = email;
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public String getComment() {
		return Comment;
	}

	public void setComment(String comment) {
		Comment = comment;
	}
}
